requirements.txt
streamlit==1.28.0
requests==2.31.0
google-generativeai==0.3.2
cloudinary==1.40.0
python-dotenv==1.0.0